requirements.txt
pytest==6.2.4
python-dotenv==0.18.0
requests==2.25.1
